task2/requirements.txt
numpy
pillow
tqdm
matplotlib